backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
pydantic
aiosqlite
numpy
faiss-cpu
PyMuPDF
pypdf
tenacity
ypy-websocket
y-py

